move out some common code
